Multi-channel ADC working
