--- zephyr-apps/mppt/boards/blackpill_f411ce.overlay
@@ -1,5 +1,0 @@
-/ {
-	zephyr,user {
-		io-channels = <&adc1 0>;
-	};
-};
