+++ zephyr-apps/mppt/prj.conf
@@ -4,3 +4,4 @@ CONFIG_LOG=y
 CONFIG_SERIAL=y
 CONFIG_ADC=y
 CONFIG_ADC_ASYNC=y
+
