+++ zephyr-apps/mppt/src/adc.h
@@ -1,0 +1,24 @@
+#include <zephyr/zephyr.h>
+#include <zephyr/drivers/adc.h>
+
+// Find devicetree configurations
+#if !DT_NODE_EXISTS(DT_PATH(zephyr_user)) || \
+	!DT_NODE_HAS_PROP(DT_PATH(zephyr_user), io_channels)
+#error "No suitable devicetree overlay specified"
+#endif
+
+#define DT_SPEC_AND_COMMA(node_id, prop, idx) \
+	ADC_DT_SPEC_GET_BY_IDX(node_id, idx),
+
+#define LABEL_AND_COMMA(node_id, prop, idx) \
+	DT_LABEL(DT_IO_CHANNELS_CTLR_BY_IDX(node_id, idx)),
+
+// ADC Settings Todo: Move this to the overlay file
+#define ADC_RESOLUTION		12
+#define ADC_GAIN		ADC_GAIN_1
+#define ADC_REFERENCE		ADC_REF_INTERNAL
+#define ADC_ACQUISITION_TIME	ADC_ACQ_TIME_DEFAULT
+
+void setup_adc();
+
+void poll_adc_channels();
