+++ zephyr-apps/overlays/blackpill_f411ce.overlay
@@ -3,13 +3,21 @@
 / {
     zephyr,user {
         signal-gpios = <&gpioc 14 GPIO_ACTIVE_HIGH>;
-		io-channels = <&adc1 4>;
-    };
 
+		/* 
+		Specify the channels to use, corresponds to ADC0-9
+		which are pins PA0-7, PB0, PB1
+		*/ 
+		io-channels = <&adc1 1>, <&adc1 4>;
+    };
 };
 
+// Update configuration from board dts file
 &adc1 {
-	pinctrl-0 = <&adc1_in4_pa4>;
+	
+	// Define each pin on the board to be used for ADC here
+	pinctrl-0 = <&adc1_in1_pa1>, <&adc1_in4_pa4>;
+	
 	pinctrl-names = "default";
 	status = "okay";
 };
